*move MergeOtherAndQuantizeLinear, MergeQuantizedConvolutionAndQuantizedActivation, MergeQuantizedAdd, MergeQuantizedMul, MergeQuantizedHardSigmoid, MergeQuantizedHswish, MergeQuantizedPrelu, MergeQuantizedScale, MergeQuantizedShuffleV0, MergeThreeQuantizedConvolutions, MergeTwoQuantizedConvolutions, SkipUnnecessaryDequantizeQuantizeV0, SkipUnnecessaryDequantizeQuantizeV1 and SkipUnnecessaryDequantize to separate source files.
