+++ prj/vs2022/CvtCore.vcxproj
@@ -38,6 +38,7 @@
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeInnerProductAndScale.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeMish.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeNormalize.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeOtherAndQuantizeLinear.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeParallelConvolutions.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeParallelDepthwiseConvolutions.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeParallelScaleAndDepthwiseConvolution.cpp" />
@@ -45,7 +46,14 @@
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergePooling.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergePowerAndScaleAndPower.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergePrelu.cpp" />
-    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantized.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedAdd.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedConvolutionAndQuantizedActivation.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedHardSigmoid.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedHswish.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedMul.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedPrelu.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedScale.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedShuffle.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedSqueezeExcitation.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeRnnGruBd.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeScale.cpp" />
@@ -55,14 +63,18 @@
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSqueezeExcitation.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSwish.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeThreeConvolutions.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeThreeQuantizedConvolutions.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeTiledScale2D.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeTwoConvolutions.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeTwoQuantizedConvolutions.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeUnpack.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeYoloV7.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\Optimizer.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\ReduceTensorIteratorIO.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\SimplifyInterp.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\SkipTwoPermutes.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\SkipUnnecessaryDequantize.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\SkipUnnecessaryDequantizeQuantize.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\TransposeConvolutions.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\TransposeInnerProduct.cpp" />
     <ClInclude Include="..\..\src\Cvt\Common\Params.h" />
