+++ src/Cvt/Optimizer/SkipUnnecessaryDequantize.cpp
@@ -1,0 +1,47 @@
+/*
+* Synet Framework (http://github.com/ermig1979/Synet).
+*
+* Copyright (c) 2018-2025 Yermalayeu Ihar.
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in
+* all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*/
+
+#include "Cvt/Optimizer/Common.h"
+#include "Cvt/Optimizer/Optimizer.h"
+
+namespace Synet
+{
+    bool SkipUnnecessaryDequantize(const LayerParams& src, size_t& index, QuantizationMethod method, LayerParams& dst, Changes& changes)
+    {
+        if (src.size() < index + 2)
+            return false;
+        const LayerParam& dl = src[index + 0];
+        LayerParam layer = src[index + 1];
+        if (dl.type() != LayerTypeDequantizeLinear)
+            return false;
+        if (layer.type() != LayerTypeMeta && layer.meta().type() != MetaTypeShape)
+            return false;
+        if (InsideLink(src, index, 2))
+            return false;
+        layer.src() = dl.src();
+        dst.push_back(layer);
+        index += 1;
+        return true;
+    }
+}
